reporting_mvp/requirements.txt
Django==2.1.3
requests==2.20.1
psycopg2-binary==2.7.6.1
social-auth-app-django==2.1.0
pika==0.12.0
dramatiq==1.3.0
django-dramatiq==0.5.0
asana==0.8.2
psycopg2==2.7.5
redis==2.10.6
django-configurations==2.1
djangorestframework==3.9.0
django-filter==2.0.0
Markdown==3.0.1
